Crate pallet_msa

MSA Pallet

The MSA pallet provides functionality for handling Message Source Accounts.

Config

Call

Pallet

Overview

The MSA pallet provides functions for:

Creating, reading, updating, and deleting operations for MSAs.

Managing delegation relationships for MSAs.

Managing keys associated with MSA.

Terminology

MSA - Message Source Account. A Source or Provider Account for MRC Messages. It may or may not have Capacity token. It must have at least one AccountId (public key) associated with it. Created by generating a new MSA ID number and associating it with a Substrate AccountID. An MSA is required for sending Capacity-based messages and for creating Delegations.

MSA ID - This is the ID number created for a new Message Source Account and associated with a Substrate AccountId.

Delegator - a Message Source Account that has provably delegated certain actions to a Provider, typically sending a Message

Provider - the actor that a Delegator has delegated specific actions to.

Delegation - A stored Delegator-Provider association between MSAs which permits the Provider to perform specific actions on the Delegator’s behalf.

Implementations

The MSA pallet implements the following traits:

AccountProvider: Functions for accessing and validating MSAs. This implementation is what is used in the runtime.

Interface

Dispatchable Functions

add_key_to_msa - Associates a key to an MSA ID in a signed payload.

add_provider_to_msa - Creates a delegation relationship between a Provider and MSA.

create - Creates an MSA for the Origin.

create_sponsored_account_with_delegation - Origin creates an account for a given AccountId and sets themselves as a Provider.

remove_delegation_by_provider - Provider MSA terminates a Delegation with Delegator MSA by expiring it.

revoke_msa_delegation_by_delegator - Delegator MSA terminates a Delegation with the Provider MSA by expiring it.

revoke_msa_key - Revokes the given key by expiring it.

Assumptions

Total MSA keys should be less than the constant Config::MSA::MaxKeys.

Re-exports

pub use pallet::*;

pub use types::AddKeyData;

pub use types::AddProvider;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

types

Types for the MSA Pallet

weights

Autogenerated weights for pallet_msa

Functions

wrap_binary_data

Wraps PREFIX and POSTFIX around a Vec<u8> Returns PREFIX ++ data ++ POSTFIX

Type Definitions

MessageSourceId

Message Source Id or msaId is the unique identifier for Message Source Accounts