Trait pallet_msa::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type ConvertIntoAccountId32: Convert<Self::AccountId, AccountId32>;
    type MaxPublicKeysPerMsa: Get<u8>;
    type MaxSchemaGrantsPerDelegation: Get<u32> + Clone + Debug + Eq;
    type MaxProviderNameSize: Get<u32>;
    type SchemaValidator: SchemaValidator<SchemaId>;
    type MortalityWindowSize: Get<u32>;
    type MaxSignaturesPerBucket: Get<u32>;
    type NumberOfBuckets: Get<u32>;
    type MaxSignaturesStored: Get<Option<u32>>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type ConvertIntoAccountId32: Convert<Self::AccountId, AccountId32>

AccountId truncated to 32 bytes

type MaxPublicKeysPerMsa: Get<u8>

Maximum count of keys allowed per MSA

type MaxSchemaGrantsPerDelegation: Get<u32> + Clone + Debug + Eq

Maximum count of schemas granted for publishing data per Provider

type MaxProviderNameSize: Get<u32>

Maximum provider name size allowed per MSA association

type SchemaValidator: SchemaValidator<SchemaId>

A type that will supply schema related information.

type MortalityWindowSize: Get<u32>

The number of blocks per virtual “bucket” in the PayloadSignatureRegistry Virtual buckets are the first part of the double key in the PayloadSignatureRegistry StorageDoubleMap. This permits a key grouping that enables mass removal of stale signatures which are no longer at risk of replay.

type MaxSignaturesPerBucket: Get<u32>

The maximum number of signatures that can be assigned to a virtual bucket. In other words, no more than this many signatures can be assigned a specific first-key value.

type NumberOfBuckets: Get<u32>

The total number of virtual buckets There are exactly NumberOfBuckets first-key values in PayloadSignatureRegistry.

type MaxSignaturesStored: Get<Option<u32>>

The maximum number of signatures that can be stored in PayloadSignatureRegistry. This MUST be MaxSignaturesPerBucket * NumberOfBuckets. It’s separate because this config is provided to signature storage and cannot be a calculated value.