Module pallet_msa::types

Types for the MSA Pallet

Structs

AddKeyData

A type definition for the payload of adding an MSA key - pallet_msa::add_public_key_to_msa

AddProvider

Structure that is signed for granting permissions to a Provider

Delegation

Struct for the information of the relationship between an MSA and a Provider

DelegatorId

A DelegatorId an MSA Id serving the role of a Delegator. Delegators delegate to Providers. Encodes and Decodes as just a u64

KeyInfoResponse

RPC Response for getting getting MSA ids

ProviderId

Provider is the recipient of a delegation. It is a subset of an MSA Encodes and Decodes as just a u64

Constants

EMPTY_FUNCTION

Dispatch Empty

Traits

PermittedDelegationSchemas

The interface for mutating schemas permissions in a delegation relationship.

Type Definitions

MessageSourceId

Message Source Id or msaId is the unique identifier for Message Source Accounts