Trait pallet_msa::types::PermittedDelegationSchemas

pub trait PermittedDelegationSchemas<T: Config> {
    fn try_insert_schema(
        &mut self,
        schema_id: SchemaId
    ) -> Result<(), DispatchError>;
    fn try_get_mut_schema(
        &mut self,
        schema_id: SchemaId,
        block_number: T::BlockNumber
    ) -> Result<(), DispatchError>;

    fn try_insert_schemas(
        &mut self,
        schema_ids: Vec<SchemaId>
    ) -> Result<(), DispatchError> { ... }
    fn try_get_mut_schemas(
        &mut self,
        schema_ids: Vec<SchemaId>,
        block_number: T::BlockNumber
    ) -> Result<(), DispatchError> { ... }
}

The interface for mutating schemas permissions in a delegation relationship.

Required Methods

fn try_insert_schema(&mut self, schema_id: SchemaId) -> Result<(), DispatchError>

Attempt to insert a new schema. Dispatches error when the max allowed schemas are exceeded.

fn try_get_mut_schema(
&mut self,
schema_id: SchemaId,
block_number: T::BlockNumber
) -> Result<(), DispatchError>

Attempt get and mutate a schema. Dispatches error when a schema cannot be found.

Provided Methods

fn try_insert_schemas(
&mut self,
schema_ids: Vec<SchemaId>
) -> Result<(), DispatchError>

Attempt to insert a collection of schemas. Dispatches error when the max allowed schemas are exceeded.

fn try_get_mut_schemas(
&mut self,
schema_ids: Vec<SchemaId>,
block_number: T::BlockNumber
) -> Result<(), DispatchError>

Attempt get and mutate a collection of schemas. Dispatches error when a schema cannot be found.